backend/requirements.txt
dashscope>=1.23.6
Pillow>=11.0.0
python-dotenv>=1.0.0
pdf2image>=1.17.0